threads_refresh_token/requirements.txt
functions-framework==3.*
requests==2.*
google-cloud-secret-manager==2.* 